docker/requirements.txt
fire
float_range
jupyter
librosa
matplotlib
notebook
numpy
pandas
pillow
pyvirtualdisplay
pyscreenshot
selenium
scipy
scikit-learn
